Rework folder structure and Makefile. TODO: Fix implementation of operation table
