+++ main.c
@@ -6,7 +6,7 @@
 /*   By: sgoffaux <sgoffaux@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/14 13:26:41 by sgoffaux          #+#    #+#             */
-/*   Updated: 2021/09/21 14:47:38 by sgoffaux         ###   ########.fr       */
+/*   Updated: 2021/09/21 16:30:39 by sgoffaux         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,6 @@ int	main(int argc, char **argv, char **envp)
 	{	
 		if (parse(&script))
 			continue ;
-		// execution
 		free_commands(&script);
 	}
 	system("leaks minishell");
