+++ srcs/tokenizer.c
@@ -1,0 +1,41 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   tokenizer.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: sgoffaux <sgoffaux@student.s19.be>         +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/09/21 14:54:35 by sgoffaux          #+#    #+#             */
+/*   Updated: 2021/09/21 16:18:19 by sgoffaux         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+int	tokenizer(char *str, t_token **head)
+{
+	t_operations	curr;
+	char			*prev;
+
+	prev = str;
+	while (str && *str)
+	{
+		curr = search_token_type(str);
+		if ((curr.op != 0 || *str == '\"' || *str == '\'') && prev != str)
+			add_token(head, create_token(prev, str - prev, TOKEN_NAME));
+		if (curr.op != 0)
+		{
+			str += curr.size;
+			if (curr.type != TOKEN_EAT)
+				add_token(head, create_token(curr.op, curr.size, curr.type));
+			prev = str;
+		}
+		else if ((*str == '\"' || *str == '\'') && !treat_quotes(&str))
+			return (0);
+		else
+			++str;
+	}
+	if (prev != str)
+		add_token(head, create_token(prev, str - prev, TOKEN_NAME));
+	return (1);
+}
